# [openrr-planner] Use AABB-based pruning in environmental collision checking