This PR implements an AABB (Axis Aligned Bounding Box)-based sweep and prune algorithm for environmental collision checking, as same as #630 .

We can evaluate the improvement in computational time with the following command:

```bash
cargo bench bench_check_environmental_collisions
```

The comparison result in my environment is:
```
[Before]
bench_check_environmental_collisions                        
                        time:   [362.01 us 362.96 us 363.89 us]

[After (this PR is applied)]
bench_check_environmental_collisions                        
                        time:   [12.597 us 12.633 us 12.670 us]
                        change: [-96.604% -96.553% -96.508%] (p = 0.00 < 0.05)
                        Performance has improved.
```